requirements.txt
streamlit
langchain
langchain-openai
langchain-community
langchain-text-splitters
openai
faiss-cpu
pypdf
python-dotenv
tiktoken
